moved product_state into prod_simulator
